Cleanup based on PR comments
