+++ lib/assertions.js
@@ -5,8 +5,9 @@ import isChecked from './assertions/is-checked';
 import isNotChecked from './assertions/is-not-checked';
 import isRequired from './assertions/is-required';
 import isNotRequired from './assertions/is-not-required';
+import isVisible from './assertions/is-visible';
+import isNotVisible from './assertions/is-not-visible';
 
-import visible from './helpers/visible';
 import elementToString from './helpers/element-to-string';
 import collapseWhitespace from './helpers/collapse-whitespace';
 
@@ -135,6 +136,34 @@ export default class DOMAssertions {
    */
   isNotRequired(message) {
     isNotRequired.call(this, message);
+  }
+
+  /**
+   * Assert that the [HTMLElement][] or an [HTMLElement][] matching the
+   * `selector` is visible.
+   *
+   * @param {string?} message
+   *
+   * @example
+   * assert.dom('.foo').isVisible();
+   *
+   */
+  isVisible(message) {
+    isVisible.call(this, message);
+  }
+
+  /**
+   *  Assert that the [HTMLElement][] or an [HTMLElement][] matching the
+   * `selector` is not visible.
+   *
+   * @param {string?} message
+   *
+   * @example
+   * assert.dom('.foo').isNotVisible();
+   *
+   */
+  isNotVisible(message) {
+    isNotVisible.call(this, message);
   }
 
   /**
@@ -573,58 +602,6 @@ export default class DOMAssertions {
 
   lacksValue(message) {
     this.hasNoValue(message);
-  }
-
-  /**
-   * Assert an [HTMLElement][] matching the `selector` is visible.
-   *
-   * @param {string?} message
-   *
-   * @example
-   * assert.dom('.foo').isVisible();
-   *
-   */
-  isVisible(message) {
-    let element = this.findTargetElement();
-    if (!element) return;
-
-    let result = visible(element);
-    let actual = result
-      ? `Element ${this.target} is visible`
-      : `Element ${this.target} is not visible`;
-    let expected = `Element ${this.target} is visible`;
-
-    if (!message) {
-      message = expected;
-    }
-
-    this.pushResult({ result, actual, expected, message });
-  }
-
-  /**
-   * Assert an [HTMLElement][] matching the `selector` is not visible.
-   *
-   * @param {string?} message
-   *
-   * @example
-   * assert.dom('.foo').isNotVisible();
-   *
-   */
-  isNotVisible(message) {
-    let element = this.findTargetElement();
-    if (!element) return;
-
-    let result = !visible(element);
-    let actual = result
-      ? `Element ${this.target} is not visible`
-      : `Element ${this.target} is visible`;
-    let expected = `Element ${this.target} is not visible`;
-
-    if (!message) {
-      message = expected;
-    }
-
-    this.pushResult({ result, actual, expected, message });
   }
 
   /**
